requirements.txt
# data science
numpy<2.0.0
pandas==2.1.4

# Uncomment if you use sklearn
scikit-learn==1.5.2
scipy==1.13.1

# Install the correct TensorFlow version
# Uncomment if you use TensorFlow
tensorflow-cpu==2.16.1    # For both Intel and Apple silicon
keras==3.7.0

# Google Cloud
# Uncomment if you use Google Cloud Storage
# google-cloud-storage

# API
fastapi
uvicorn

# Seb
nltk
gensim==4.3.3
joblib
#transformers
#xgboost
#torch
